# Avoid unnecessary allocations in PriorityQueue.EnqueueRange